test(plugins): keep reconciler coverage within lint budget
